Schema/find

+++ src/util/impl/ByteReader.tpp
@@ -108,7 +108,7 @@ void ByteReader<Datatype>::Impl::read(Numeric& item) {
 
   item = 0;
   for (uint32_t i = 0; i < item_size; ++i) {
-    item |= static_cast<unsigned char>(bytes_[offset_] << (i * 8));
+    item |= static_cast<Numeric>(static_cast<unsigned char>(bytes_[offset_]) << (i * 8));
     ++offset_;
   }
 }
